package refactor, dockerize other modules
