+++ ascii-generator/src/main/docker/Dockerfile
@@ -1,0 +1,6 @@
+FROM java:8
+VOLUME /tmp
+ADD ascii-generator-1.0.0-SNAPSHOT.jar /app.jar
+RUN bash -c 'touch /app.jar'
+EXPOSE 8080
+ENTRYPOINT ["java","-Djava.security.egd=file:/dev/./urandom","-jar","/app.jar"]
