+++ ascii-generator/src/main/java/com/norasoft/generator/AsciiGeneratorApplication.java
@@ -1,4 +1,4 @@
-package pl.com.bms;
+package com.norasoft.generator;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
