+++ ascii-generator/src/main/java/com/norasoft/generator/generator/ConversionException.java
@@ -1,4 +1,4 @@
-package pl.com.bms.generator;
+package com.norasoft.generator.generator;
 
 public class ConversionException extends RuntimeException {
