+++ ascii-generator/src/test/groovy/com/norasoft/generator/generator/Ascii.groovy
@@ -1,4 +1,4 @@
-package pl.com.bms.generator
+package com.norasoft.generator.generator
 
 class Ascii {
     static def expectedAscii() {
