+++ eureka/src/main/java/com/norasoft/eureka/Eureka.java
@@ -1,4 +1,4 @@
-package pl.com.bms;
+package com.norasoft.eureka;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
