+++ picture-repo/src/main/java/com/norasoft/picturerepo/PictureRepoApplication.java
@@ -1,4 +1,4 @@
-package pl.com.bms;
+package com.norasoft.picturerepo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
